phase-II/backend/requirements.txt
# Production Dependencies for Todo Backend
# Updated for Vercel serverless deployment

# Core Framework
fastapi>=0.115.0,<0.116.0
mangum>=0.17.0,<1.0.0  # ASGI adapter for Vercel/AWS Lambda

# Database
sqlmodel>=0.0.22,<0.1.0
asyncpg>=0.29.0,<1.0.0
psycopg2-binary>=2.9.9,<3.0.0  # Required for Alembic migrations
alembic>=1.14.0,<2.0.0

# Data Validation
pydantic>=2.10.0,<3.0.0
email-validator>=2.1.0,<3.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0

# Rate Limiting
slowapi>=0.1.9,<1.0.0

# Server
uvicorn[standard]>=0.32.0,<1.0.0

# Environment
python-dotenv>=1.0.0,<2.0.0
